requirements.txt
# Core Scientific Libraries
pandas>=1.3.0
numpy>=1.20.0

# Optimization
jmetalpy>=1.5.5

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bars
tqdm>=4.60.0

# Data Formats
openpyxl>=3.0.0
pyarrow>=6.0.0

# Configuration
PyYAML>=5.4.0

# Presentation Generation (optional)
python-pptx>=0.6.21

# Development Dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Type Hints
types-PyYAML>=6.0.0

# Web API and Database (Bucha Virtual)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyodbc>=4.0.39

# Machine Learning - Deep Learning (Anomaly Detection)
torch>=2.0.0
scikit-learn>=1.3.0
